refact(user): Integrate user settings with develop.
