+++ lib/features/user/view/screens/settings_screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:telware_cross_platform/core/theme/dimensions.dart';
 import 'package:telware_cross_platform/core/theme/palette.dart';
-import 'package:telware_cross_platform/features/auth/view/widget/profile_header_widget.dart';
-import 'package:telware_cross_platform/features/auth/view/widget/settings_option_widget.dart';
-import 'package:telware_cross_platform/features/auth/view/widget/settings_section.dart';
+import 'package:telware_cross_platform/features/user/view/widget/profile_header_widget.dart';
+import 'package:telware_cross_platform/features/user/view/widget/settings_option_widget.dart';
+import 'package:telware_cross_platform/features/user/view/widget/settings_section.dart';
 
 class SettingsScreen extends StatefulWidget {
   static const String route = '/settings';
@@ -27,14 +27,16 @@ class _SettingsScreen extends State<SettingsScreen> {
       "title": "Account",
       "options": [
         {
+          "key": "change-number-option",
           "text": user["phoneNumber"],
           "subtext": "Tap to change phone number",
           "routes": "/change-number"
         },
-        {"text": user["username"], "subtext": "Username"},
+        {"text": "@${user["username"]}", "subtext": "Username"},
         {
-          "text": "Bio",
-          "subtext": "Add a few words about yourself",
+          "key": "bio-option",
+          "text": user["bio"] ?? "Bio",
+          "subtext": user["bio"] != null ? "Bio" : "Add a few words about yourself",
           "routes": "/bio"
         }
       ]
@@ -45,9 +47,10 @@ class _SettingsScreen extends State<SettingsScreen> {
         {
           "icon": Icons.chat_bubble_outline,
           "text": 'Chat Settings',
-          "routes": '/chat-settings'
+          "routes": 'locked'
         },
         {
+          "key": "privacy-option",
           "icon": Icons.lock_outline_rounded,
           "text": 'Privacy and Security',
           "routes": '/privacy'
@@ -55,29 +58,29 @@ class _SettingsScreen extends State<SettingsScreen> {
         {
           "icon": Icons.notifications_none,
           "text": 'Notifications and Sounds',
-          "routes": '/notifications'
+          "routes": 'locked'
         },
         {
           "icon": Icons.pie_chart_outline,
           "text": 'Data and Storage',
-          "routes": '/data-storage'
+          "routes": 'locked'
         },
         {
           "icon": Icons.battery_saver_outlined,
           "text": 'Power Saving',
-          "routes": '/power-saving'
+          "routes": 'locked'
         },
         {
           "icon": Icons.folder_outlined,
           "text": 'Chat Folders',
-          "routes": '/chat-folders'
+          "routes": 'locked'
         },
-        {"icon": Icons.devices, "text": 'Devices', "routes": '/devices'},
+        {"icon": Icons.devices, "text": 'Devices', "routes": 'locked'},
         {
           "icon": Icons.language_rounded,
           "text": 'Language',
           "trailing": 'English',
-          "routes": '/language'
+          "routes": 'locked'
         },
       ]
     },
@@ -85,19 +88,20 @@ class _SettingsScreen extends State<SettingsScreen> {
       "options": [
         {
           "icon": Icons.star_border_purple500_rounded,
-          "text": 'TelWare Premium'
+          "text": 'TelWare Premium',
+          "routes": "locked",
         },
-        {"icon": Icons.star_border_rounded, "text": 'My Stars'},
-        {"icon": Icons.storefront, "text": 'TelWare Business'},
-        {"icon": Icons.card_giftcard, "text": 'Send a Gift'},
+        {"icon": Icons.star_border_rounded, "text": 'My Stars', "routes": "locked"},
+        {"icon": Icons.storefront, "text": 'TelWare Business', "routes": "locked"},
+        {"icon": Icons.card_giftcard, "text": 'Send a Gift', "routes": "locked"},
       ]
     },
     const {
       "title": "Help",
       "options": [
-        {"icon": Icons.chat, "text": 'Ask a Question'},
-        {"icon": Icons.question_mark, "text": 'TelWare FAQ'},
-        {"icon": Icons.verified_user_outlined, "text": 'Privacy Policy'},
+        {"icon": Icons.chat, "text": 'Ask a Question', "routes": "locked"},
+        {"icon": Icons.question_mark, "text": 'TelWare FAQ', "routes": "locked"},
+        {"icon": Icons.verified_user_outlined, "text": 'Privacy Policy', "routes": "locked"},
       ],
       "trailing": "TalWare for Android v11.2.0 (5299) store bundled arm64-v8a"
     }
@@ -115,7 +119,7 @@ class _SettingsScreen extends State<SettingsScreen> {
             pinned: true,
             leading: IconButton(
                 icon: const Icon(Icons.arrow_back),
-                onPressed: () => Navigator.pushNamed(context, "/profile")),
+                onPressed: () => Navigator.pop(context)),
             actions: const [
               Icon(Icons.search),
               SizedBox(width: 16),
@@ -143,6 +147,7 @@ class _SettingsScreen extends State<SettingsScreen> {
                 settingsOptions: [],
                 actions: [
                   SettingsOptionWidget(
+                    key: ValueKey("set-profile-photo-option"),
                     icon: Icons.camera_alt_outlined,
                     iconColor: Palette.primary,
                     text: "Set Profile Photo",
