+++ lib/main.dart
@@ -1,18 +1,23 @@
 import 'package:flutter/material.dart';
-import 'core/theme/auth_theme.dart';
-import 'features/auth/view/screens/user_profile_screen.dart';
+import 'package:telware_cross_platform/features/user/view/screens/block_user.dart';
+import 'package:telware_cross_platform/features/auth/view/screens/change_number_form_screen.dart';
+import 'package:telware_cross_platform/features/user/view/screens/blocked_users.dart';
+import 'package:telware_cross_platform/features/user/view/screens/change_number_screen.dart';
+import 'package:telware_cross_platform/features/user/view/screens/privacy_and_security_screen.dart';
+import 'package:telware_cross_platform/features/user/view/screens/profile_info_screen.dart';
+import 'package:telware_cross_platform/features/user/view/screens/settings_screen.dart';
+import 'package:telware_cross_platform/core/theme/app_theme.dart';
+import 'package:telware_cross_platform/features/user/view/screens/user_profile_screen.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:telware_cross_platform/features/stories/models/contact_model.dart';
 import 'package:hive_flutter/hive_flutter.dart';
-// import 'core/theme/auth_theme.dart';
-import 'features/stories/models/story_model.dart';
-import 'features/home/view/screen/home_screen.dart';
+import 'package:telware_cross_platform/features/stories/models/story_model.dart';
+import 'package:telware_cross_platform/features/home/view/screens/home_screen.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:telware_cross_platform/core/view/screen/splash_screen.dart';
 import 'package:telware_cross_platform/features/auth/view/screens/log_in_screen.dart';
 import 'package:telware_cross_platform/features/auth/view_model/auth_view_model.dart';
-import 'core/models/user_model.dart';
-import 'core/theme/app_theme.dart';
+import 'package:telware_cross_platform/core/models/user_model.dart';
 import 'features/auth/view/screens/sign_up_screen.dart';
 import 'features/auth/view/screens/verification_screen.dart';
 
@@ -61,8 +66,16 @@ class _TelWareState extends ConsumerState<TelWare> {
         SignUpScreen.route: (context) => const SignUpScreen(),
         VerificationScreen.route: (context) => const VerificationScreen(),
         HomeScreen.route: (context) => const HomeScreen(),
+        SettingsScreen.route: (context) => const SettingsScreen(),
+        ChangeNumberScreen.route : (context) => const ChangeNumberScreen(),
+        ChangeNumberFormScreen.route: (context) => const ChangeNumberFormScreen(),
+        ProfileInfoScreen.route: (context) => const ProfileInfoScreen(),
+        BlockUserScreen.route: (context) => const BlockUserScreen(),
+        BlockedUsersScreen.route: (context) => const BlockedUsersScreen(),
+        UserProfileScreen.route: (context) => const UserProfileScreen(),
+        PrivacySettingsScreen.route: (context) => const PrivacySettingsScreen(),
       },
-      initialRoute: SplashScreen.route,
+      initialRoute: HomeScreen.route,
     );
   }
 }
